+++ api/todo/controllers/trash.controller.ts
@@ -9,38 +9,42 @@ export default class TrashController {
         const trashed = await TrashService.GetTrash(email);
 
         if (trashed) res.status(200).send(trashed);
+        else res.status(404).send();
       } catch (err: any) {
         console.log(err);
+        res.status(500).send(err.toString());
       }
     } else res.status(500).send();
   }
 
   public static async RestoreItem(req: Request, res: Response): Promise<void> {
     const email = req.params.email;
-    if (email) {
+    const listId = req.params.listId;
+    if (email && listId) {
       try {
-        const listId = req.params.listId;
         const response = await TrashService.RestoreItem(email, listId);
 
         if (response.succeeded) res.status(204).send();
         else res.status(422).send(response.message);
       } catch (err: any) {
         console.log(err);
+        res.status(500).send();
       }
     } else res.status(500).send();
   }
 
   public static async DeleteItem(req: Request, res: Response): Promise<void> {
     const email = req.params.email;
-    if (email) {
+    const listId = req.params.listId;
+    if (email && listId) {
       try {
-        const listId = req.params.listId;
         const response = await TrashService.DeleteItem(email, listId);
 
         if (response.succeeded) res.status(204).send();
         else res.status(422).send(response.message);
       } catch (err: any) {
         console.log(err);
+        res.status(500).send();
       }
     } else res.status(500).send();
   }
@@ -55,6 +59,7 @@ export default class TrashController {
         else res.status(422).send(response.message);
       } catch (err: any) {
         console.log(err);
+        res.status(500).send(err.toString());
       }
     } else res.status(500).send();
   }
